+++ packages/simple-form-joy/src/form-components/form-checkbox/tests/form-checkbox.spec.tsx
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/experimental-ct-react";
 
-import { alwaysErrorValidatorMessage } from "src/test-components/form-utils";
+import { alwaysErrorValidatorMessage } from "src/test-components/use-test-form";
 
 import { FormCheckboxTestComponent } from "./form-checkbox-test-component";
