+++ packages/simple-form-joy/src/form-components/form-switch/tests/form-switch.spec.tsx
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/experimental-ct-react";
 
-import { alwaysErrorValidatorMessage } from "src/test-components/form-utils";
+import { alwaysErrorValidatorMessage } from "src/test-components/use-test-form";
 
 import { FormSwitchTestComponent } from "./form-switch-test-component";
