+++ packages/simple-form-joy/src/form-components/form-text/tests/form-text.spec.tsx
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/experimental-ct-react";
 
-import { alwaysErrorValidatorMessage } from "src/test-components/form-utils";
+import { alwaysErrorValidatorMessage } from "src/test-components/use-test-form";
 
 import { FormTextTestComponent } from "./form-text-test-component";
